% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disgust_cb}
\alias{disgust_cb}
\title{Three domains of disgust codebook.}
\format{A data frame with descriptions for each of the 26 variables in
  the \code{disgust} dataset: \code{id}, \code{user_id},
  \code{start}, \code{end} and 21 disgust questions (7 each from
  \strong{moral}, \strong{sexual}, and \strong{pathogen} domains).}
\usage{
disgust_cb
}
\description{
Codebook for \code{disgust} dataset.
Three domain disgust scale from Tybur, J. M., Lieberman, D. L., &
Griskevicius, V. G. (2009). Microbes, mating, and morality: Individual
differences in three functional domains of disgust.
\emph{Journal of Personality and Social Psychology}, 29, 103-122.
doi: 10.1037/a0015474
}
\keyword{datasets}
