% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faceatt}
\alias{faceatt}
\title{Face Attractiveness Ratings.}
\format{A data frame with 2513 observations and 19 variables:
  \code{rater_sex}, \code{rater_sexpref}, \code{rater_age}, \code{end} and
  102 face attractiveness ratings (\code{X001} to \code{X173}).}
\usage{
faceatt
}
\description{
Facial attractiveness ratings of 102 adult faces (info in \code{faceinfo}) from
DeBruine, Lisa; Jones, Benedict (2017): Face Research Lab London Set.
\emph{figshare}. doi: 10.6084/m9.figshare.5047666
}
\details{
Raters (n = 2513) rated 102 neutral front-facing faces on a 1-7 scale from
"much less attractiveness than average" to "much more attractive than average".
}
\keyword{datasets}
