% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faceinfo}
\alias{faceinfo}
\title{Face Information.}
\format{A data frame with 102 observations and 4 variables:
  \code{face_id}, \code{face_age}, \code{face_sex}, and \code{face_sex}.}
\usage{
faceinfo
}
\description{
Demographic information about 102 adult faces from
DeBruine, Lisa; Jones, Benedict (2017): Face Research Lab London Set.
\emph{figshare}. doi: 10.6084/m9.figshare.5047666
}
\keyword{datasets}
