% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motivation}
\alias{motivation}
\title{Motivation.}
\format{A data frame with 909 observations and 35 variables :
  \code{a01} - \code{a10} Images of architecture from Tybur,
  \code{p01} - \code{p10} Images of pathogens from Schaller,
  \code{g01} - \code{g10} Images of gun violence from Schaller}
\usage{
motivation
}
\description{
Keypress motivation task with images of architecture, guns, and pathogens.
}
\details{
Instructions: In this experiment, you will look at images for 1.5 minutes.
How long you look at each image is up to you. You can change the time each
image is on the screen by pressing buttons like you did in the
button-pressing training task you completed at the beginning of this study.
You might find some of the images upsetting.
}
\keyword{datasets}
