% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratings_m}
\alias{ratings_m}
\title{Ratings of Male Faces and Bodies.}
\format{A data frame with 535 observations and 57 variables.}
\usage{
ratings_m
}
\description{
13 judgments of 50 male faces and bodies from
Morrison, Danielle; Wang, Hongyi; Hahn, Amanda C.; Jones, Benedict C.;
DeBruine, Lisa M. (2017). Predicting the reward value of faces and bodies
from social perception. PLoS One. doi: 10.1371/journal.pone.0185093
Supplemental Materials. doi: 10.17605/OSF.IO/G27WF https://osf.io/g27wf/
}
\keyword{datasets}
