% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senseek}
\alias{senseek}
\title{Sensation seeking scale.}
\format{A data frame with 10K observations and 19 variables:
  \code{id}, \code{user_id}, \code{start}, \code{end} and
  14 sensation seeking questions (\code{sss1} to \code{sss14}).}
\usage{
senseek
}
\description{
Sensation seeking scale from Zuckerman M. (1984). Sensation seeking:
a comparative approach to a human trait.
\emph{Behavioral and Brain Sciences}. 7: 413-471. doi: 10.1017/S0140525X00018938
The codebook is available at \code{senseek_cb}.
}
\details{
Participants are given 14 forced-choice questions, such as
\dQuote{I would like a job that requires a lot of traveling.} versus
\dQuote{I would prefer a job in one location.}. Questions were scored as 0 for
the less sensation-seeking option, and 1 for the more sensation-seeking
option.
}
\keyword{datasets}
