% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senseek_cb}
\alias{senseek_cb}
\title{Sensation seeking scale codebook.}
\format{A data frame with descriptions for each of the 19 variables in
  the \code{senseek} dataset:  \code{id}, \code{user_id},
  \code{start}, \code{end} and 14 sensation seeking questions
  (\code{sss1} to \code{sss14}).}
\usage{
senseek_cb
}
\description{
Codebook for \code{senseek} dataset.
Sensation seeking scale from Zuckerman M. (1984). Sensation seeking:
a comparative approach to a human trait.
\emph{Behavioral and Brain Sciences}. 7: 413-471. doi: 10.1017/S0140525X00018938
}
\keyword{datasets}
