% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stim_f}
\alias{stim_f}
\title{Information about Female Faces and Bodies.}
\format{A data frame with 50 observations and 9 variables:
  \code{stimulus name}, \code{stimulus sex}, \code{age}, \code{height(cm)},
  \code{weight(kg)}, \code{BMI}, \code{chest(cm)}, \code{waist(cm)},
  and \code{hips(cm)}.}
\usage{
stim_f
}
\description{
Information about 50 female faces and bodies from 3d.sk reported in
Morrison, Danielle; Wang, Hongyi; Hahn, Amanda C.; Jones, Benedict C.;
DeBruine, Lisa M. (2017). Predicting the reward value of faces and bodies
from social perception. PLoS One. doi: 10.1371/journal.pone.0185093
Supplemental Materials. doi: 10.17605/OSF.IO/G27WF https://osf.io/g27wf/
}
\details{
Ratings of these stimuli in the \code{ratings_f} dataset.
}
\keyword{datasets}
