% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{users}
\alias{users}
\title{Participant data.}
\format{A data frame with 53999 observations and 3 variables:
  \code{user_id}, \code{sex}, and \code{birthday}.}
\usage{
users
}
\description{
A dataset containing the user_id, sex and birthdate of participants in the
\code{disgust}, \code{senseek}, and \code{motivation} datasets.
Join using \code{user_id}.
}
\keyword{datasets}
